# Optimize buffered paginated store merging