--- a/ddsketch/store/collapsing_highest_dense_store.go
+++ b/ddsketch/store/collapsing_highest_dense_store.go
@@ -136,9 +136,10 @@ func (s *CollapsingHighestDenseStore) MergeWith(other Store) {
 	}
 	o, ok := other.(*CollapsingHighestDenseStore)
 	if !ok {
-		for bin := range other.Bins() {
-			s.AddBin(bin)
-		}
+		other.ForEach(func(index int, count float64) (stop bool) {
+			s.AddWithCount(index, count)
+			return false
+		})
 		return
 	}
 	if o.minIndex < s.minIndex || o.maxIndex > s.maxIndex {