--- a/ddsketch/store/collapsing_lowest_dense_store.go
+++ b/ddsketch/store/collapsing_lowest_dense_store.go
@@ -141,9 +141,10 @@ func (s *CollapsingLowestDenseStore) MergeWith(other Store) {
 	}
 	o, ok := other.(*CollapsingLowestDenseStore)
 	if !ok {
-		for bin := range other.Bins() {
-			s.AddBin(bin)
-		}
+		other.ForEach(func(index int, count float64) (stop bool) {
+			s.AddWithCount(index, count)
+			return false
+		})
 		return
 	}
 	if o.minIndex < s.minIndex || o.maxIndex > s.maxIndex {